[Lox] Reduce Allocations
Interpreter would clear the memory for every variable and before every pass through interp, it now adds an extra component to track the current runtime type of every entity. Thus we only allocate memory up to once for each type.

Fixes a memory leak that was found with long loops.